move util methods to `utils.py`
